WhatsAppBot/requirements.txt
cachetools==4.1.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
dialogflow==1.0.0
Flask==1.1.1
google-api-core==1.20.0
google-auth==1.17.2
googleapis-common-protos==1.52.0
grpcio==1.29.0
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PySocks==1.7.1
pytz==2019.3
requests==2.22.0
rsa==4.6
six==1.14.0
twilio==6.25.2
urllib3==1.25.8
Werkzeug==1.0.0